requirements.txt
django-allauth
django-markdownx
djangorestframework
openai
tailwindcss